backend/requirements.txt
flask
flask-SQLAlchemy
ipykernel
psycopg2-binary
flask-migrate
flask-moment
flask-wtf
babel
pandas
#
aniso8601
Click
Flask-Cors
Flask-RESTful
itsdangerous
Jinja2
MarkupSafe
pytz
six
SQLAlchemy
Werkzeug
python-dotenv